requirements.txt
flask
pyyaml
rdflib
munch
owlrl
markdown